+++ tests/eval_structure/test_evaluable_graph.py
@@ -9,6 +9,7 @@ from pytestarch.eval_structure.evaluable_architecture import (
 )
 from pytestarch.eval_structure.evaluable_graph import EvaluableArchitectureGraph
 from pytestarch.eval_structure_impl.networkxgraph import NetworkxGraph
+from pytestarch.exceptions import ImpossibleMatch
 from pytestarch.importer.import_types import AbsoluteImport
 
 MODULE_1 = "Module1"
@@ -34,7 +35,7 @@ MODULE_F = "E.F"  # submodule of E
 
 
 @pytest.fixture(scope="module")
-def evaluable() -> EvaluableArchitecture:
+def evaluable() -> EvaluableArchitectureGraph:
     all_modules = [MODULE_1, MODULE_2, MODULE_3, MODULE_4, SUB_MODULE_OF_2, MODULE_6]
     imports = [
         AbsoluteImport(MODULE_1, MODULE_2),
@@ -49,7 +50,7 @@ def evaluable() -> EvaluableArchitecture:
 
 
 @pytest.fixture(scope="module")
-def submodule_evaluable() -> EvaluableArchitecture:
+def submodule_evaluable() -> EvaluableArchitectureGraph:
     all_modules = [MODULE_A, MODULE_B, MODULE_C, MODULE_D, MODULE_E, MODULE_F]
     imports = [
         AbsoluteImport(MODULE_F, MODULE_C),
@@ -783,8 +784,8 @@ def test_not_other_to_module_than_between_named_modules(
 @pytest.mark.parametrize(
     "imports",
     [
-        # [AbsoluteImport(MODULE_1, SUB_MODULE_OF_2)],
-        # [AbsoluteImport(SUB_MODULE_OF_1, SUB_MODULE_OF_2)],
+        [AbsoluteImport(MODULE_1, SUB_MODULE_OF_2)],
+        [AbsoluteImport(SUB_MODULE_OF_1, SUB_MODULE_OF_2)],
         [AbsoluteImport(MODULE_3, MODULE_2)],
     ],
 )
@@ -914,3 +915,95 @@ def test_when_module_name_in_alias_spec_is_not_found_an_error_is_raised(
 
     with pytest.raises(KeyError, match=non_existent_module):
         evaluable.visualize(aliases=aliases)
+
+
+MODULE = "test.module.submodule.class"
+
+
+partial_match_test_cases = [
+    pytest.param("test.module.submodule.class", True, id="complete match - match"),
+    pytest.param("test.module.submodule.clasz", False, id="complete match - no match"),
+    pytest.param("*.class", True, id="end match - match"),
+    pytest.param("*.clasz", False, id="end match - no match"),
+    pytest.param("test.*", True, id="start match - match"),
+    pytest.param("test,*", False, id="start match - no match"),
+    pytest.param("*le.submo*", True, id="inner match - match"),
+    pytest.param("les.submo*", False, id="inner match - no match"),
+]
+
+
+@pytest.mark.parametrize("match, expected_result", partial_match_test_cases)
+def test_partial_match(match: str, expected_result: bool) -> None:
+    assert (
+        EvaluableArchitectureGraph._partial_name_match(match, MODULE) == expected_result
+    )
+
+
+def test_module_matches(submodule_evaluable: EvaluableArchitectureGraph) -> None:
+    module_matches = [
+        Module(name="*D", partial_match=True),
+        Module(name="*A*", partial_match=True),
+    ]
+
+    calculated_matches = submodule_evaluable._convert_to_full_name_matches(
+        [module_matches]
+    )
+
+    assert len(calculated_matches) == 1
+    assert len(calculated_matches[0]) == 3
+
+    assert Module(name=MODULE_A) in calculated_matches[0]
+    assert Module(name=MODULE_B) in calculated_matches[0]
+    assert Module(name=MODULE_D) in calculated_matches[0]
+
+
+def test_non_partial_module_does_not_match_anything_but_itself(
+    submodule_evaluable: EvaluableArchitectureGraph,
+) -> None:
+    module_matches = [Module(name="*.*")]
+
+    calculated_matches = submodule_evaluable._convert_to_full_name_matches(
+        [module_matches]
+    )
+
+    assert len(calculated_matches) == 1
+    assert len(calculated_matches[0]) == 1
+
+    assert Module(name="*.*") in calculated_matches[0]
+
+
+def test_never_matched_match_raises_error(
+    evaluable: EvaluableArchitectureGraph,
+) -> None:
+    module_matches = [Module(name="I will not match", partial_match=True)]
+
+    with pytest.raises(
+        ImpossibleMatch, match="No modules found that match: I will not match"
+    ):
+        evaluable._convert_to_full_name_matches([module_matches])
+
+
+def test_module_matches_reported_separately_by_group(
+    submodule_evaluable: EvaluableArchitectureGraph,
+) -> None:
+    module_matches_1 = [
+        Module(name="*D", partial_match=True),
+    ]
+
+    module_matches_2 = [
+        Module(name="*A*", partial_match=True),
+    ]
+
+    calculated_matches = submodule_evaluable._convert_to_full_name_matches(
+        [module_matches_1, module_matches_2]
+    )
+
+    assert len(calculated_matches) == 2
+
+    assert len(calculated_matches[0]) == 1
+    assert len(calculated_matches[1]) == 3
+
+    assert Module(name=MODULE_D) in calculated_matches[0]
+    assert Module(name=MODULE_A) in calculated_matches[1]
+    assert Module(name=MODULE_B) in calculated_matches[1]
+    assert Module(name=MODULE_D) in calculated_matches[1]