+++ docs/changelog.md
@@ -4,7 +4,10 @@ This project uses semantic versioning and follows [keep a changelog](https://kee
 
 
 ## [Unreleased]
+
+## [1.3.0] -- 2022-11-28
 ### Added
+- Reference to modules only by part of their names
 - Aliases for plot labels
 - Test generation from PUML component diagrams